ratesmart/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
gunicorn>=21.2,<22.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.6,<7.0
textblob>=0.18,<0.19
